+++ sshd-sftp/src/main/java/org/apache/sshd/client/subsystem/sftp/fs/SftpAclFileAttributeView.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.sshd.client.subsystem.sftp;
+package org.apache.sshd.client.subsystem.sftp.fs;
 
 import java.io.IOException;
 import java.nio.file.LinkOption;
@@ -28,6 +28,7 @@ import java.nio.file.attribute.PosixFileAttributes;
 import java.nio.file.attribute.UserPrincipal;
 import java.util.List;
 
+import org.apache.sshd.client.subsystem.sftp.SftpClient;
 import org.apache.sshd.client.subsystem.sftp.impl.AbstractSftpFileAttributeView;
 
 /**